# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Blog API",
    "description": "CSE 341 Project 2 - Express, Mongoose, Express-Validator",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/entries": {
      "get": {
        "tags": ["Entries"],
        "summary": "Retrieve a list of entries",
        "description": "Retrieve a list of entries from the database.",
        "responses": {
          "200": {
            "description": "A list of entries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": { "type": "string" },
                      "title": { "type": "string" },
                      "content": { "type": "string" },
                      "comments": { "type": "integer" },
                      "likes": { "type": "integer" },
                      "userID": { "type": "string" },
                      "created_at": { "type": "string", "format": "date" }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "List of entries not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 404 },
                    "message": { "type": "string", "example": "Entries not found" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 500 },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Entries"],
        "summary": "Create a new entry",
        "description": "Create a new entry in the database.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": { "type": "string" },
                "content": { "type": "string" },
                "comments": { "type": "array", "items": { "type": "string" } },
                "likes": { "type": "integer" },
                "userID": { "type": "string" },
                "created_at": { "type": "string", "format": "date" }
              },
              "required": ["title", "content", "userID"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Entry created.",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "entryID": { "type": "string" } } }
              }
            }
          },
          "422": {
            "description": "Error on validation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 422 },
                    "message": { "type": "string", "example": "Error on entry creation." },
                    "details": {
                      "type": "object",
                      "properties": {
                        "[prop]": { "type": "string", "example": "[Error description]" }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 500 },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/entries/{id}": {
      "get": {
        "tags": ["Entries"],
        "summary": "Retrieve an existing entry",
        "description": "Retrieve an entry from the database.",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "An existing entry.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "title": { "type": "string" },
                    "content": { "type": "string" },
                    "comments": { "type": "array", "items": { "type": "string" } },
                    "likes": { "type": "integer" },
                    "created_at": { "type": "string", "format": "date" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "_id": { "type": "string" },
                        "username": { "type": "string" },
                        "firstName": { "type": "string" },
                        "lastName": { "type": "string" },
                        "email": { "type": "string" },
                        "birthday": { "type": "string", "format": "date" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error on validation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 400 },
                    "message": { "type": "string", "example": "Error on find ID." },
                    "details": {
                      "type": "object",
                      "properties": { "id": { "type": "string", "example": "Invalid ID." } }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Entry not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 404 },
                    "message": { "type": "string", "example": "Entry not found" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 500 },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Entries"],
        "summary": "Update an existing entry",
        "description": "Update an existing entry in the database.",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": { "type": "string" },
                "content": { "type": "string" },
                "comments": { "type": "array", "items": { "type": "string" } },
                "likes": { "type": "integer" },
                "userID": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "204": { "description": "Entry updated successfully. No content in response." },
          "400": {
            "description": "Error on validation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 400 },
                    "message": { "type": "string", "example": "Error on entry creation." },
                    "details": {
                      "type": "object",
                      "properties": {
                        "[prop]": { "type": "string", "example": "[Error description]" }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 500 },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Entries"],
        "summary": "Delete an existing entry",
        "description": "Delete an entry from the database.",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Entry deleted successfully. No content in response." },
          "400": {
            "description": "Error on validation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 400 },
                    "message": { "type": "string", "example": "Error on find ID." },
                    "details": {
                      "type": "object",
                      "properties": { "id": { "type": "string", "example": "Invalid ID." } }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 500 },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
  "securityDefinitions": {},
...
}